GH-1 : Just refactoring
