--- src/main/java/com/amalvadkar/lms/auth/app/enums/ErrorMsgEnum.java
@@ -1,16 +1,0 @@
-package com.amalvadkar.lms.auth.app.enums;
-
-public enum ErrorMsgEnum {
-
-    EMAIL_NOT_EXIST("Email not exist");
-
-    String value;
-
-    private ErrorMsgEnum(String value){
-        this.value=value;
-    }
-
-    public String getValue(){
-        return value;
-    }
-}
