--- src/main/java/com/amalvadkar/lms/auth/app/exception/OtpNotValidException.java
@@ -1,9 +1,0 @@
-package com.amalvadkar.lms.auth.app.exception;
-
-import org.springframework.http.HttpStatus;
-
-public class OtpNotValidException extends AuthException {
-    public OtpNotValidException(String message) {
-        super(message, HttpStatus.NOT_FOUND.value());
-    }
-}
